+++ scripts/build.js
@@ -1,10 +1,12 @@
 const tasks = require('./tasks');
+const shell = require('shelljs');
 
 tasks.replaceWebpack();
 console.log('[Copy assets]');
 console.log('-'.repeat(80));
 tasks.copyAssets('build');
+console.log('');
 
 console.log('[Webpack Build]');
 console.log('-'.repeat(80));
-exec('webpack --config webpack/prod.config.js --progress --profile --colors');
+shell.exec('webpack --config webpack/prod.config.js --progress --profile --colors');