+++ scripts/tasks.js
@@ -1,4 +1,5 @@
-require('shelljs/global');
+const shell = require('shelljs');
+const cspBuilder = require('content-security-policy-builder');
 
 exports.replaceWebpack = () => {
   const replaceTasks = [{
@@ -9,14 +10,53 @@ exports.replaceWebpack = () => {
     to: 'node_modules/webpack-hot-middleware/process-update.js'
   }];
 
-  replaceTasks.forEach(task => cp(task.from, task.to));
+  replaceTasks.forEach(task => shell.cp(task.from, task.to));
+};
+
+const getContentSecurityPolicy = (isProd) => {
+  const directives = {
+    defaultSrc: ["'self'"],
+    scriptSrc: ["'self'"],
+    connectSrc: ["'self'", 'https://rpc.adshares.net', 'https://rpc.e11.click', 'data:'],
+    styleSrc: ['*', "'unsafe-inline'"],
+    fontSrc: ["'self'", 'https://fonts.gstatic.com'],
+    imgSrc: ["'self'", 'data:'],
+  };
+
+  if (!isProd) {
+    directives.scriptSrc.push('http://localhost:3000', 'https://localhost:3000', "'unsafe-eval'");
+    directives.connectSrc.push('http://localhost:3000', 'https://localhost:3000', 'http://127.0.0.12:5000');
+    directives.styleSrc.push('blob:');
+  }
+
+  return { directives };
 };
 
 exports.copyAssets = (type) => {
-  const env = type === 'build' ? 'prod' : type;
-  rm('-rf', type);
-  mkdir(type);
-  cp(`chrome/manifest.${env}.json`, `${type}/manifest.json`);
-  cp('-R', 'chrome/assets/*', type);
-  exec(`pug -O "{ env: '${env}' }" -o ${type} chrome/views/`);
+  const isProd = type === 'build';
+  const view = JSON.stringify({
+    env: isProd ? 'prod' : type,
+    name: process.env.npm_package_display_name,
+    version: process.env.npm_package_version,
+    description: process.env.npm_package_description,
+    author: `${process.env.npm_package_author_name} <${process.env.npm_package_author_email}>`,
+    host: isProd ? '' : 'http://localhost:3000',
+    csp: cspBuilder(getContentSecurityPolicy(isProd))
+  });
+
+  shell.rm('-rf', type);
+  shell.mkdir(type);
+  console.log(`- created ${type} directory`);
+
+  shell.cp('-R', 'chrome/assets/*', type);
+  console.log(`- copied chrome/assets into ${type}`);
+
+  shell.exec(`echo ${view} | mustache - chrome/manifest.mustache ${type}/manifest.json`);
+  console.log(`- rendered ${type}/manifest.json`);
+
+  shell.ls('chrome/views/*.mustache').forEach((file) => {
+    const output = `${type}/${file.replace(/^.*[\\/]/, '').replace(/[^.]+$/, 'html')}`;
+    shell.exec(`echo ${view} | mustache - ${file} ${output}`);
+    console.log(`- rendered ${output}`);
+  });
 };