+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Password min length info on restore page
+- Build process
 
 # [0.2.1] - 2019-01-16
 ### Added